burundi_au_chairmanship/backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-jazzmin>=2.6
Pillow>=10.0
